Fix regression bug that caused the recorded urls not be decoded.
Unfortenately there was no test for urls that look different after encoding. Now there is, grrr.

It kind of feels bad to mutate the raw (encoded) url in .lastWithUrlDecoded... One option would be to add a new entry with a key 'decodedUrl' but in the mean time there is no use case where the original url (e.g. '%3Cfoo%3E' instead of encoded '<foo>') would be needed. Therefore sticking with the mutation.